+++ src/org/schema/constants/traits/PlanAction.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace org\schema\constants\traits;
+
+/**
+ * The PlanAction properties enumeration.
+ */
+trait PlanAction
+{
+    const string SCHEDULED_TIME = 'scheduledTime' ;
+}